requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
scipy>=1.11.0
scikit-surprise>=1.1.3
flask>=3.0.0
jinja2>=3.1.0
gunicorn>=21.2.0
streamlit>=1.34.0
matplotlib>=3.8.0
seaborn>=0.13.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
